+++ Nanook/SurfaceAgency.swift
@@ -14,8 +14,8 @@ struct SurfaceAgency {
     // MARK: Properties
     
     let agency: Int
-    let frequency: Float?
-    let duration: Float?
+    let frequency: Double?
+    let duration: Double?
     let operatingDays: Int?
     let lineNames: [String]?
     let lineCodes: [String]?
@@ -26,8 +26,8 @@ struct SurfaceAgency {
 
     init(surfaceAgency: [String:AnyObject]) {
         self.agency = surfaceAgency[Constants.SurfaceAgency.agency] as! Int
-        self.frequency = surfaceAgency[Constants.SurfaceAgency.frequency] as? Float
-        self.duration = surfaceAgency[Constants.SurfaceAgency.duration] as? Float
+        self.frequency = surfaceAgency[Constants.SurfaceAgency.frequency] as? Double
+        self.duration = surfaceAgency[Constants.SurfaceAgency.duration] as? Double
         self.operatingDays = surfaceAgency[Constants.SurfaceAgency.operatingDays] as? Int
         self.lineNames = surfaceAgency[Constants.SurfaceAgency.lineNames] as? [String]
         self.lineCodes = surfaceAgency[Constants.SurfaceAgency.lineCodes] as? [String]
