+++ Nanook/SurfaceStop.swift
@@ -14,16 +14,16 @@ struct SurfaceStop {
     // MARK: Properties
     
     let place: Int
-    let transitDuration: Float
-    let stopDuration: Float
+    let transitDuration: Double
+    let stopDuration: Double
 
     
     // MARK: Initializers
     
     init(surfaceStop: [String:AnyObject]) {
         self.place = surfaceStop[Constants.SurfaceStop.place] as! Int
-        self.transitDuration = surfaceStop[Constants.SurfaceStop.transitDuration] as! Float
-        self.stopDuration = surfaceStop[Constants.SurfaceStop.stopDuration] as! Float
+        self.transitDuration = surfaceStop[Constants.SurfaceStop.transitDuration] as! Double
+        self.stopDuration = surfaceStop[Constants.SurfaceStop.stopDuration] as! Double
     }
 
 }
